+++ src/JuMP.jl
@@ -31,7 +31,7 @@ export
     getNumVars, getNumConstraints, getObjectiveValue, getObjective,
     getObjectiveSense, setObjectiveSense, writeLP, writeMPS, setObjective,
     addConstraint, addVar, addVars, addSOS1, addSOS2, solve, copy,
-    getInternalModel,
+    getInternalModel, setPresolve,
     # Variable
     setName, getName, setLower, setUpper, getLower, getUpper,
     getValue, setValue, getDual,
@@ -87,6 +87,9 @@ type Model
 
     # JuMPDict list
     dictList::Vector
+
+    # presolve callback function
+    presolve
     # storage vector for merging duplicate terms
     indexedVector::IndexedVector{Float64}
 
@@ -106,7 +109,7 @@ function Model(;solver=nothing)
         Model(QuadExpr(),:Min,LinearConstraint[], QuadConstraint[],SOSConstraint[],
               0,String[],Float64[],Float64[],Int[],
               0,Float64[],Float64[],Float64[],nothing,MathProgBase.MissingSolver("",Symbol[]),false,
-              nothing,nothing,nothing,JuMPDict[],IndexedVector(Float64,0),nothing,Dict{Symbol,Any}())
+              nothing,nothing,nothing,JuMPDict[],nothing,IndexedVector(Float64,0),nothing,Dict{Symbol,Any}())
     else
         if !isa(solver,AbstractMathProgSolver)
             error("solver argument ($solver) must be an AbstractMathProgSolver")
@@ -115,7 +118,7 @@ function Model(;solver=nothing)
         Model(QuadExpr(),:Min,LinearConstraint[], QuadConstraint[],SOSConstraint[],
               0,String[],Float64[],Float64[],Int[],
               0,Float64[],Float64[],Float64[],nothing,solver,false,
-              nothing,nothing,nothing,JuMPDict[],IndexedVector(Float64,0),nothing,Dict{Symbol,Any}())
+              nothing,nothing,nothing,JuMPDict[],nothing,IndexedVector(Float64,0),nothing,Dict{Symbol,Any}())
     end
 end
 
@@ -167,6 +170,8 @@ function copy(source::Model)
 end
 
 getInternalModel(m::Model) = m.internalModel
+
+setPresolve(m::Model, f::Function) = (m.presolve = f)
 
 ###############################################################################
 # Variable class