+++ src/solvers.jl
@@ -192,6 +192,7 @@ function solveLP(m::Model; load_model_only=false, suppress_warnings=false)
     end 
 
     if !load_model_only
+        (m.presolve != nothing) && m.presolve(m)
         optimize!(m.internalModel)
         stat = status(m.internalModel)
     else
@@ -283,6 +284,7 @@ function solveMIP(m::Model; load_model_only=false, suppress_warnings=false)
     end
 
     if !load_model_only
+        (m.presolve != nothing) && m.presolve(m)
         optimize!(m.internalModel)
         stat = status(m.internalModel)
     else